services/ai-core/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
fastapi
uvicorn
pydantic
scikit-learn
langchain-openai
numpy
langchain
openai
pgvector
psycopg2-binary
python-dotenv
requests
